Ещё классы для работы с плейлистами

+++ app/src/main/java/com/il76/playlistmaker/media/data/db/PlaylistTrackEntity.kt
@@ -23,6 +23,6 @@ import com.il76.playlistmaker.history.data.db.TrackEntity
     ]
 )
 data class PlaylistTrackEntity(
-    val playlistId: Long,
-    val trackId: Long,
+    val playlistId: Int,
+    val trackId: Int,
 )
